Changed some number types, make the migration last.
And some rustfmt updates.

+++ migrations/2018-09-05-065252_places_osm/up.sql
@@ -1,5 +1,5 @@
 -- Add fields for OpenStreetMap-based locations to places table.
-ALTER TABLE places ADD COLUMN osm_id INTEGER UNIQUE;
+ALTER TABLE places ADD COLUMN osm_id BIGINT UNIQUE;
 ALTER TABLE places ADD COLUMN osm_level SMALLINT;
 
 CREATE INDEX places_osm_idx ON places (osm_id);
